fix bugs 1.1
